backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
fake-useragent==1.4.0
flask==3.0.0
flask-cors==4.0.0
pandas>=2.1.0
numpy>=1.26.0
openai==1.3.0
httpx==0.25.0
pyairtable==3.2.0
msal==1.25.0
cloudinary==1.36.0
apscheduler==3.10.4
